requirements.txt
aioredis==1.3.1
alembic==1.6.5
appdirs==1.4.4
asgiref==3.3.4
async-timeout==3.0.1
bcrypt==3.2.0
black==21.6b0
cffi==1.14.6
click==8.0.1
cryptography==3.4.7
ecdsa==0.17.0
fastapi==0.65.2
funcy==1.16
greenlet==1.1.0
h11==0.12.0
hiredis==2.0.0
httptools==0.2.0
Mako==1.1.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.8.1
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.17.1
python-editor==1.0.4
python-jose==3.3.0
PyYAML==5.4.1
redis==3.5.3
regex==2021.4.4
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.3
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
uvloop==0.15.2
watchgod==0.7
websockets==9.1
python-multipart
pymongo
motor
pyspark
